Update to create custom datastore
