+++ go.mod
@@ -7,8 +7,10 @@ require (
 	github.com/aws/aws-sdk-go-v2/config v1.8.0
 	github.com/aws/aws-sdk-go-v2/feature/s3/manager v1.5.1
 	github.com/aws/aws-sdk-go-v2/service/s3 v1.15.0
+	github.com/benbjohnson/clock v1.3.0 // indirect
 	github.com/bsipos/thist v1.0.0
-	github.com/btcsuite/btcd v0.21.0-beta
+	github.com/btcsuite/btcd v0.22.0-beta
+	github.com/cespare/xxhash/v2 v2.1.2 // indirect
 	github.com/dapperlabs/testingdock v0.4.2
 	github.com/davecgh/go-spew v1.1.1
 	github.com/dgraph-io/badger/v2 v2.2007.3
@@ -28,33 +30,36 @@ require (
 	github.com/grpc-ecosystem/go-grpc-middleware/v2 v2.0.0-20200501113911-9a95f0fdbfea
 	github.com/grpc-ecosystem/go-grpc-prometheus v1.2.0
 	github.com/grpc-ecosystem/grpc-gateway/v2 v2.6.0
+	github.com/hashicorp/errwrap v1.1.0 // indirect
 	github.com/hashicorp/go-multierror v1.1.1
 	github.com/hashicorp/golang-lru v0.5.4
 	github.com/improbable-eng/grpc-web v0.12.0
-	github.com/ipfs/go-bitswap v0.4.0
+	github.com/ipfs/go-bitswap v0.5.1
 	github.com/ipfs/go-block-format v0.0.3
-	github.com/ipfs/go-cid v0.0.7
-	github.com/ipfs/go-datastore v0.4.6
-	github.com/ipfs/go-ds-badger2 v0.1.1
-	github.com/ipfs/go-ipfs-blockstore v0.1.6
-	github.com/ipfs/go-ipfs-ds-help v0.1.1
-	github.com/ipfs/go-ipfs-exchange-interface v0.0.1
+	github.com/ipfs/go-cid v0.1.0
+	github.com/ipfs/go-datastore v0.5.0
+	github.com/ipfs/go-ds-badger2 v0.1.2-0.20211119002906-7318f1b76158
+	github.com/ipfs/go-ipfs-blockstore v1.1.0
+	github.com/ipfs/go-ipfs-ds-help v1.1.0 // indirect
+	github.com/ipfs/go-ipfs-exchange-interface v0.1.0
 	github.com/ipfs/go-log v1.0.5
-	github.com/ipld/go-ipld-prime v0.12.0 // indirect
+	github.com/ipld/go-ipld-prime v0.14.0 // indirect
 	github.com/jrick/bitset v1.0.0
-	github.com/kr/text v0.2.0 // indirect
+	github.com/klauspost/compress v1.13.6 // indirect
 	github.com/libp2p/go-addr-util v0.1.0
-	github.com/libp2p/go-libp2p v0.14.4
-	github.com/libp2p/go-libp2p-core v0.8.6
-	github.com/libp2p/go-libp2p-discovery v0.5.0
-	github.com/libp2p/go-libp2p-kad-dht v0.13.0
-	github.com/libp2p/go-libp2p-pubsub v0.4.1
-	github.com/libp2p/go-libp2p-swarm v0.5.3
-	github.com/libp2p/go-libp2p-tls v0.1.3
-	github.com/libp2p/go-libp2p-transport-upgrader v0.4.6
-	github.com/libp2p/go-tcp-transport v0.2.7
+	github.com/libp2p/go-libp2p v0.16.0-dev.0.20211116110622-8df360043e98
+	github.com/libp2p/go-libp2p-core v0.11.0
+	github.com/libp2p/go-libp2p-discovery v0.6.0
+	github.com/libp2p/go-libp2p-kad-dht v0.15.0
+	github.com/libp2p/go-libp2p-pubsub v0.6.0
+	github.com/libp2p/go-libp2p-swarm v0.8.0
+	github.com/libp2p/go-libp2p-tls v0.3.0
+	github.com/libp2p/go-libp2p-transport-upgrader v0.5.0
+	github.com/libp2p/go-tcp-transport v0.4.0
 	github.com/m4ksio/wal v1.0.0
-	github.com/multiformats/go-multiaddr v0.3.3
+	github.com/mattn/go-isatty v0.0.14 // indirect
+	github.com/multiformats/go-base32 v0.0.4 // indirect
+	github.com/multiformats/go-multiaddr v0.4.1
 	github.com/multiformats/go-multiaddr-dns v0.3.1
 	github.com/onflow/cadence v0.19.1-0.20210920215340-75c54c6c01eb
 	github.com/onflow/flow v0.2.3-0.20211103155021-9254a66368d5
@@ -64,11 +69,13 @@ require (
 	github.com/onflow/flow-go-sdk v0.21.0
 	github.com/onflow/flow-go/crypto v0.21.3
 	github.com/onflow/flow/protobuf/go/flow v0.2.3
+	github.com/onsi/ginkgo v1.16.5 // indirect
 	github.com/opentracing/opentracing-go v1.2.0
 	github.com/pelletier/go-toml v1.9.4 // indirect
 	github.com/pierrec/lz4 v2.6.1+incompatible
 	github.com/pkg/errors v0.9.1
-	github.com/prometheus/client_golang v1.10.0
+	github.com/prometheus/client_golang v1.11.0
+	github.com/prometheus/common v0.32.1 // indirect
 	github.com/psiemens/sconfig v0.1.0 // indirect
 	github.com/rs/zerolog v1.19.0
 	github.com/sethvargo/go-retry v0.1.0
@@ -82,14 +89,17 @@ require (
 	github.com/uber/jaeger-lib v2.4.0+incompatible // indirect
 	github.com/vmihailenco/msgpack v4.0.4+incompatible
 	github.com/vmihailenco/msgpack/v4 v4.3.11
-	go.uber.org/atomic v1.7.0
-	go.uber.org/multierr v1.6.0
-	golang.org/x/crypto v0.0.0-20210513164829-c07d793c2f9a
+	go.uber.org/atomic v1.9.0
+	go.uber.org/multierr v1.7.0
+	go.uber.org/zap v1.19.1 // indirect
+	golang.org/x/crypto v0.0.0-20211117183948-ae814b36b871
 	golang.org/x/exp v0.0.0-20200224162631-6cc2880d07d6
+	golang.org/x/mod v0.5.1 // indirect
+	golang.org/x/net v0.0.0-20211118161319-6a13c67c3ce4 // indirect
 	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
-	golang.org/x/sys v0.0.0-20210910150752-751e447fb3d0
-	golang.org/x/text v0.3.7 // indirect
+	golang.org/x/sys v0.0.0-20211117180635-dee7805ff2e1
 	golang.org/x/time v0.0.0-20191024005414-555d28b269f0
+	golang.org/x/tools v0.1.7 // indirect
 	google.golang.org/api v0.49.0
 	google.golang.org/genproto v0.0.0-20210903162649-d08c68adba83
 	google.golang.org/grpc v1.40.0
@@ -97,6 +107,7 @@ require (
 	google.golang.org/protobuf v1.27.1
 	gopkg.in/ini.v1 v1.63.0 // indirect
 	gotest.tools v2.2.0+incompatible
+	lukechampine.com/blake3 v1.1.7 // indirect
 	pgregory.net/rapid v0.4.7
 )
